Gameplay update Changed how the game increases the ball speed throughout the game. Before the ball would speed up by 50 after every level. This wasn't the best design as the levels are dramatically different. If the first level only had 10 bricks, the second level would be played at a faster speed while the score would be low. Compared to if the first level had 30 bricks, the score would be much higher for less difficulty. This created a scenario where the player relied heavily on getting lucky with the big levels early on to get a high score, not to mention having these big levels early with a slow ball makes the game a boring slog.
The new design speeds up the ball every time it hits the paddle. This makes high scores less reliant on lucky level generation and makes the difficulty of each level more consistent. Also this makes the long levels more interesting as it isn't a static speed the whole level, only increasing by beating it.
It also promotes skill expression and strategy, as minimizing the number of paddle hits is key to keeping the ball speed to a manageable rate.

This new design also stresses the importance of using the powerup system to snipe out of place blocks to quickly end the level or to carve a path for the ball to follow.

Overall this change makes each level feel fun to play rather than having to grind out a bunch of tutorial feeling levels before getting to a decent speed.

As the speed increases dramatically every level, it was rare to get past level 2 before the ball moved so fast the paddle could no longer keep up with it. To combat this, the ball speed is halved at the beginning of each level. This allows some ramp up time in each level while maintaining the gradual difficulty scaling.

May need to adjust further if a player is able to get back to back short levels and rng themselvs into actually decreasing the overall ballspeed as the game progresses, but further data is needed.